Move maskEmail util
